sale endpoint

+++ src/main/java/ma/nemo/assignment/domain/Product.java
@@ -1,7 +1,6 @@
 package ma.nemo.assignment.domain;
 
 import jakarta.persistence.CascadeType;
-import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
@@ -12,7 +11,6 @@ import jakarta.persistence.Temporal;
 import jakarta.persistence.TemporalType;
 import java.util.Date;
 
-import org.hibernate.annotations.Cascade;
 
 import lombok.ToString;
 
@@ -28,7 +26,7 @@ public class Product {
   @ManyToOne(cascade = CascadeType.PERSIST)
   private ProductModel productModel;
 
-  private Integer quantity;
+  private long quantity;
 
   @Temporal(TemporalType.TIMESTAMP)
   private Date expirationDate;
@@ -57,11 +55,11 @@ public class Product {
     this.productModel = productModel;
   }
 
-  public Integer getQuantity() {
+  public long getQuantity() {
     return quantity;
   }
 
-  public void setQuantity(Integer quantity) {
+  public void setQuantity(long quantity) {
     this.quantity = quantity;
   }
